Move domino_is_reachable to helpers.py
This function was written as a helper in one test file, but it could
be generally imported and used by others if we move it to helpers.py.

Testing:

Unit tests that know about this function still pass after the move.

tests/test_basic_auth.py::test_object_creation_with_api_key PASSED
tests/test_basic_auth.py::test_object_creation_with_token_file PASSED
tests/test_basic_auth.py::test_auth_against_real_deployment_with_api_key PASSED
tests/test_basic_auth.py::test_auth_against_real_deployment_with_token_file PASSED
